8300-Refactor code DatebaseChanges
